requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
boto3==1.40.8
botocore==1.40.8
Brotli==1.1.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
Django==5.0.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.59.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
qrcode==8.2
razorpay==1.4.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.4
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.4.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
